+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Fix Metadata loading in multiple components and looping when failing #147
 - Fix BlockTxs API call happening 2x #128
 
+### Improvements
+
+- Create a filter dropdown in tx messages to filter by tx type (must have 2+ to show) #127
+
 ## 2.0.2
 
 ### Bug Fixes